rework DigitalActuator
