+++ src/plugins/spark/features/DigitalActuator/getters.ts
@@ -1,7 +1,1 @@
 export const typeName = 'DigitalActuator';
-
-export const channel = [
-  'Unset',
-  'A',
-  'B',
-];
