+++ src/plugins/spark/features/Pid/PidFull.vue
@@ -62,6 +62,7 @@ export default class PidFull extends BlockCrudComponent {
 <template>
   <q-card dark v-bind="$attrs">
     <slot name="toolbar" />
+    <slot name="warnings" />
 
     <q-card-section>
       <BlockEnableToggle
